\name{towData}
\alias{towData}
\docType{package}

\title{
Continuous Plankton Recorder (CPR) Data Analyser
}

\description{
This package is designed for the purpose of data analysis and visualisation for the Mediterranean Continuous Plankton Recorder Survey (MedCPR) project.
}

\author{
\enumerate{
\item Kamil Erguler, Associate Research Scientist, EEWRC, The Cyprus Institute \email{k.erguler@cyi.ac.cy}
}
}

\keyword{cpr}
\keyword{plankton}
\keyword{data}


